Fix structures being detected but not visible if scanned
A structure could lose its sighted status under a scan if a marine was
around, because in that case the marine would have been the "last viewer".
Therefor superseeding the scan SetSighted() call. Now the scan is registering
itself as a viewer, and a XZ distance check is done to ensure it is seing
anything in its range. If an Ink cloud is triggered, then as usual the
structure loses its viewer and sighted status.

This change will also allow ARCs not to lose track of buildings if manually
ordered to attack.